SignInForm fix redirect, Profile redesign
