+++ src/app/profile/page.tsx
@@ -1,4 +1,5 @@
 import Container from "@/components/Container/Container";
+import Page from "@/components/Page/Page";
 import Typography from "@/components/Typography/Typography";
 import { authConfig } from "@/configs/auth";
 import { getServerSession } from "next-auth";
@@ -8,22 +9,32 @@ async function Profile() {
   const session = await getServerSession(authConfig);
 
   return (
-    <main className="flex  flex-col justify-between  max-w-[1440px]">
-      <Container className=" gap-5">
-        <Image
-          src={session?.user?.image ? session.user.image : "/google-square.svg"}
-          alt={session?.user?.name ? session.user.name : "user-avatar"}
-          width={96}
-          height={96}
-        />
-        <Typography size="32px" tag="h2">
-          {session?.user?.email}
-        </Typography>
-        <Typography size="24px" tag="h2">
-          {session?.user?.name}
-        </Typography>
+    <Page>
+      <Container className=" gap-5 justify-start">
+        <div className=" flex justify-between w-full">
+          <Typography size="24px" tag="h2">
+            Wellcome {session?.user?.name}
+          </Typography>
+          <div className=" flex gap-4 items-top">
+            <div className=" rounded-full overflow-hidden self-center">
+              <Image
+                src={
+                  session?.user?.image
+                    ? session.user.image
+                    : "/google-square.svg"
+                }
+                alt={session?.user?.name ? session.user.name : "user-avatar"}
+                width={22}
+                height={22}
+              />
+            </div>
+            <Typography size="20px" tag="h2">
+              {session?.user?.email}
+            </Typography>
+          </div>
+        </div>
       </Container>
-    </main>
+    </Page>
   );
 }
